requirements.txt
torch>=2.2
torchvision>=0.17
transformers>=4.39
numpy>=1.26
tqdm>=4.66
accelerate>=0.30
scikit-learn>=1.4
Pillow>=10.0
python-dotenv>=1.0
pillow
